Revert "Core/Misc: Update money from 32 bit to 64 bit, increasing maximum money to 1 milion (cataclysm value)"
This reverts commit 229c278cbd4395be2cfe90c7042dda86e0806e04.

Needs more testing!

+++ src/server/game/Chat/Commands/Level3.cpp
@@ -7121,7 +7121,7 @@ bool ChatHandler::HandleSendMoneyCommand(const char *args)
         return false;
 
     char* money_str = strtok(NULL, "");
-    int64 money = money_str ? atoll(money_str) : 0;
+    int32 money = money_str ? atoi(money_str) : 0;
     if (money <= 0)
         return false;
